Convert WorkflowPool to an interface.
